1. added three modes to the transfer  a. ability to lock down so only the project-broker address can transfer  b. default: the owner and an approved address set by the owner can transfer  c. combination of the above. In addition the centralised-broker addresses form a ban list - if set the transfer will fail for these addresses
2. mint-token-twenty
takes up to 20 hashes and metadata urls and cycles through delegating to the original mint-token to do the actual work.
Exits at the first metadata url thats less then 10 chars - i.e. allow the ui to pass any number from 1 to 20 inclusive and it still works

3. minting
the two mint methods now also take a clientMintPrice - allows the mint price to be set in the client - but drops to the floor mint price set on at the contract level if the client cheats by passing in a v. small mint price. This is the ultimate flexibility in determining mint price and follows teh strategy for the beneficiaries.

the client can also now pass in the buyNowPrice - value of 0 is the old / default behaviour and means the owner has to call set-sale-data before selling the NFT. A value greater than 0 puts the item up for Buy Now straight away.